requirements.txt
mudforge
esper
kdtree
rtree
dataclasses-json
orjson
django
django-idmap
ruamel.yaml
django-yamlfield
argon2-cffi
